fixed a circular dependency
