+++ app/api/templates/specs/templates.spec.js
@@ -81,7 +81,7 @@ describe('templates', () => {
     describe('when updateng a property label', () => {
       it('should update the name and update all documents using this template', (done) => {
         let newTemplate = {name: 'created_template', properties: [ {label: 'label 1'}, {label: 'label 2'}]};
-        spyOn(documents, 'updateMetadataProperties').and.returnValue(new Promise((resolve) => resolve()));
+        spyOn(templates, 'updateMetadataProperties').and.returnValue(new Promise((resolve) => resolve()));
         templates.save(newTemplate)
         .then((createdTemplate) => getTemplate(createdTemplate.id))
         .then((template) => {
@@ -90,7 +90,7 @@ describe('templates', () => {
           return templates.save(template);
         })
         .then((updatedTemplate) => {
-          expect(documents.updateMetadataProperties).toHaveBeenCalledWith(updatedTemplate.id, {
+          expect(templates.updateMetadataProperties).toHaveBeenCalledWith(updatedTemplate.id, {
             label_1: 'new_label_1',
             label_2: 'new_label_2'
           }, []);
@@ -103,7 +103,7 @@ describe('templates', () => {
     describe('when removing properties', () => {
       it('should remove the properties on all documents using the template', (done) => {
         let newTemplate = {name: 'created_template', properties: [ {label: 'label 1'}, {label: 'label 2'}, {label: 'label 3'}]};
-        spyOn(documents, 'updateMetadataProperties').and.returnValue(new Promise((resolve) => resolve()));
+        spyOn(templates, 'updateMetadataProperties').and.returnValue(new Promise((resolve) => resolve()));
         templates.save(newTemplate)
         .then((createdTemplate) => getTemplate(createdTemplate.id))
         .then((template) => {
@@ -111,7 +111,7 @@ describe('templates', () => {
           return templates.save(template);
         })
         .then((updatedTemplate) => {
-          expect(documents.updateMetadataProperties).toHaveBeenCalledWith(updatedTemplate.id, {}, ['label_1', 'label_3']);
+          expect(templates.updateMetadataProperties).toHaveBeenCalledWith(updatedTemplate.id, {}, ['label_1', 'label_3']);
           done();
         })
         .catch(done.fail);
@@ -181,9 +181,29 @@ describe('templates', () => {
     });
   });
 
+  describe('countByTemplate', () => {
+    it('should return how many documents using the template passed', (done) => {
+      templates.countByTemplate('template1')
+      .then((count) => {
+        expect(count).toBe(2);
+        done();
+      })
+      .catch(done.fail);
+    });
+
+    it('should return 0 when no count found', (done) => {
+      templates.countByTemplate('newTemplate')
+      .then((count) => {
+        expect(count).toBe(0);
+        done();
+      })
+      .catch(done.fail);
+    });
+  });
+
   describe('delete', () => {
     it('should delete a template when no document is using it', (done) => {
-      spyOn(documents, 'countByTemplate').and.returnValue(Promise.resolve(0));
+      spyOn(templates, 'countByTemplate').and.returnValue(Promise.resolve(0));
       request.get(dbURL + '/c08ef2532f0bd008ac5174b45e033c93')
       .then(template => {
         return templates.delete(template.json);
@@ -202,7 +222,7 @@ describe('templates', () => {
     });
 
     it('should throw an error when there is documents using it', (done) => {
-      spyOn(documents, 'countByTemplate').and.returnValue(Promise.resolve(1));
+      spyOn(templates, 'countByTemplate').and.returnValue(Promise.resolve(1));
       request.get(dbURL + '/c08ef2532f0bd008ac5174b45e033c93')
       .then(template => {
         return templates.delete(template.json);
@@ -235,6 +255,50 @@ describe('templates', () => {
         done();
       })
       .catch(catchErrors(done));
+    });
+  });
+
+  describe('updateMetadataProperties', () => {
+    let getDocumentsByTemplate = (template) => request.get(dbURL + '/_design/documents/_view/metadata_by_template?key="' + template + '"')
+    .then((response) => {
+      return response.json.rows.map((r) => r.value);
+    });
+
+    it('should update metadata property names on the documents matching the template', (done) => {
+      let nameChanges = {property1: 'new_name1', property2: 'new_name2'};
+      templates.updateMetadataProperties('template1', nameChanges)
+      .then(() => getDocumentsByTemplate('template1'))
+      .then((docs) => {
+        expect(docs[0].metadata.new_name1).toBe('value1');
+        expect(docs[0].metadata.new_name2).toBe('value2');
+        expect(docs[0].metadata.property3).toBe('value3');
+
+        expect(docs[1].metadata.new_name1).toBe('value1');
+        expect(docs[1].metadata.new_name2).toBe('value2');
+        expect(docs[1].metadata.property3).toBe('value3');
+        done();
+      })
+      .catch(done.fail);
+    });
+
+    it('should delete properties passed', (done) => {
+      let nameChanges = {property2: 'new_name'};
+      let deleteProperties = ['property1', 'property3'];
+      templates.updateMetadataProperties('template1', nameChanges, deleteProperties)
+      .then(() => getDocumentsByTemplate('template1'))
+      .then((docs) => {
+        expect(docs[0].metadata.property1).not.toBeDefined();
+        expect(docs[0].metadata.new_name).toBe('value2');
+        expect(docs[0].metadata.property2).not.toBeDefined();
+        expect(docs[0].metadata.property3).not.toBeDefined();
+
+        expect(docs[1].metadata.property1).not.toBeDefined();
+        expect(docs[1].metadata.new_name).toBe('value2');
+        expect(docs[1].metadata.property2).not.toBeDefined();
+        expect(docs[1].metadata.property3).not.toBeDefined();
+        done();
+      })
+      .catch(done.fail);
     });
   });
